Remove failing test
The parser is able to handle datetimes without a trailing Z - it interprets
them as valid datetimes in the local timezone.

As far as I can tell, this is valid according to the spec:

> You may omit the local offset and let the parser or host application decide
> that information. A good default is to use the host machine's local offset.

We unfortunately can't move this to the `valid` folder, since the current
implementation parses in the user's timezone, whatever timezone that is, so
whatever value got hardcoded in the `.json` file would fail for every user not
using that timezone.

Another alternative would be to modify the parser to disallow timestamps that
don't have a timezone set, or always default to UTC; I'm OK with anything but
I think it would be good to fix the tests.