Add flag "-skip" to skip some tests
Some tests on the TOML library are failing because this toml-test
library added some tests for stuff that's not supported yet. This should
be fixed, of course, but can't fix everything in a day and in the
meanwhile always having a certain subset of tests fail is somewhat
annoying, makes it easy to miss *new* test failures, and potentially
confusing for contributors.

So add a -skip flag to skip those tests until they're fixed.

Also add the test that's being run in the output:

	toml-test toml-test-decoder: 128 passed, 0 failed, 4 skipped
	toml-test toml-test-encoder:  61 passed, 0 failed, 6 skipped

Instead of just:

	128 passed, 0 failed, 4 skipped
	61 passed, 0 failed, 6 skipped

Also adds a go.mod, because otherwise it's hard/annoying to build this
in recent Go versions.