Split out command to ./cmd/tomltest; various small improvements
Split out the test runner and actual command to tomltest and
./cmd/tomltest packages. The practical upshot of this is that you can
call the rest runner directly from your tests; removing the need to
install an external binary, and you should get some code coverage for it
as well. While only useful for Go libraries, it's a nice thing for the
toml tests and won't hurt anyone else.

Rewrite the CLI to use my little zgo.at/zli package; the flag package in
Go is finicky and even the standard "go" commands contain all sorts of
hacks to work around the problems. For example you can't have flags
after positional arguments, and adding "-v" a chore rather than
<Up>v<Enter>.

This fixes #9 by passing taking all of os.Args[1:] as the parser-cmd:

	$ toml-encoder my-decoder my-decoder-arg
	$ toml-encoder -- my-decoder -x  # -- to stop flag parsing.

List the tests to run with -run=.. – there wasn't really any good way to
fix it otherwise; quoting could work (toml-test 'my-decoder -x'), but
shell quoting/word splitting is hairy and I'd rather not inflict that
upon people (see: the original space-separated -tags for the go command,
could be really hard to script with that).

Also accept glob patterns for -run and -skip.

Further improve the output and expand the docs a bit, and a few other
minor things.