Allow people to pass their own Parser, add more tests
This makes the tests easier to run for the toml library, and also allows
some other things like testing the specific error messages and meta.

Also add a whole bunch of tests extracted from the TOML library.

This adds ".multi" files to invalid tests; because you only want to use
a single test failure, you're going to end up with a lot
"integer-*.toml" tests, and managing these is a pain. So the
gen-multi.py will read these and generate a test for every line.

It was easier and less complex to do it like this than adding direct
support to the test runner, and also a lot less chance of accidental
bugs in there.