Test whitespace has higher precedence than date-time sep
In TOML, the separator between a date and time can be ` `, `T`, `t`.  A
malformed parser might consume trailing ` ` and assume a time will be
following, erroring when its anything else.  We ran into this with
`toml_edit` (ordian/toml_edit#269) because our parser is LL(1) (no
backtracing) by default and we have to opt-in to LL(k) parsing
(backtrack when time isn't present).

Fixes #97