Add test case for mixed array where the first element is an inline table
This tests a case that some libraries might have where they will check
if the first entry is a table/map/hash/assoc array and then encode it as
a table array. This was a reasonable thing to do before TOML 1.0 since
arrays could only contain one type, but now it's no longer.