Add -json, remove -testdir
Add -json flag to output as JSON.

Remove -testdir support; this is from before the tests were embedded in
the binary. I didn't want to remove it back then to retain some
compatibility, but it's been 4 years. AFAIK no one is using it and I
can't really think of a reasonable use case.