Run both decoder and encoder in one toml-test invocation
This changes the CLI so you can run both the decoder and encoder with
one command:

	toml-test -decoder="foo" -encoder="bar"

I never really liked how you had to run toml-test twice, once for the
"decoder" and once for the "encoder". It was a bit awkward but okay.
However with setup-toml-test and toml-test-matrix it's much more of a
pain and I need to merge the reports, which is not so obvious to do well
since both "valid" and "encoder" tests are both just "valid" tests.

So just change it. The encoder tests are now as "encoder/foo" instead of
"valid/foo" (for e.g. -skip and -run).

---

This is not a compatible change; however the "master" branch is renamed
to "main", and the "master" branch is kept around for compatibility.
This means people fetching from "master" will continue to work.

The next version will not be compatible anyway as the default will be
TOML 1.1 rather than 1.0.