Consider 60 minutes in a TZ offset invalid
The reason it allowed `:60` minutes in the TZ offset is because
https://github.com/golang/go/issues/67470 – in the commit that ended up
fixing it[1], a comment of "The range test use > rather than >= as some
people do write offsets of 24 hours or 60 minutes or 60 seconds" was
added. I don't know who "some people" are though.

I didn't want to be *too* pedantic, so I copied that.

I can't find any files that do that though – at least not on GitHub[2] –
so it's probably okay to up the pedantry?

[1]: https://github.com/golangclub/go/commit/1849ce6a45640ec4a6e63138211eac4276473437
[2]: https://github.com/search?q=%2F%5B%2B-%5D%5Cd%5Cd%3A60%2F+path%3A*.toml&type=code

Fixes #169