Remove "list" command
Originally added to allow people to copy test for a specific TOML
version (see: #134), but the "copy" command is now a better way to do
it.

The reason I'd like to remove it is that people keep modifying these
generated files in PRs, which is the wrong way to do it. I also had some
amount of trouble writing a good usage text explaining why this command
is useful, which is often a good indication that it's not.

So I don't see much point in retaining it? If someone really wants it,
then it can be re-add I suppose. Or copy can be improved to suit the use
case.